TIC TAC TOE bugfix2 - sign out
Removes the board from the screen when you sign out and reset board
sign in brings board back to same state but allows you to have a create
new game button. Committing change to this point and implementing check
for state of create button on sign in. I think this will solve the
problem and fix sign out issue.