TIC TAC TOE bugfix2 - sign out
Tried several hide methods for the create game button. Decided to just
remove the sucker altogether  - Ties create game function to successful
sign on and actually allows for log out and back in to exact state of
game. I find this to be a win. Just a few more rough UI edges to work
out and I do feel confident in the state of the ui.