TIC TAC TOE bugfix3
Greatly improves flow and fixes issues with sign in sign out further.
Automatically updates the number of games completed at sign on and at
game over scenario - further removing another element from the page.

Need to ensure reset button is there after signing back in as you cannot
reset a game if it is completed when you sign back in.