TIC TAC TOE bugfix3 - sign out / change password
Successfully catches if succeeds or fails and alerts accordingly. Site
no longer breaks if you sign out and then try to reset password.
Considering further logic but for now moving on. Next step is to
remove the need for a reset game button which also introduces a bug
if you log out at the end of a game. When you log back in the game is
still going but cannot be reset.