TIC TAC TOE master
Found a bug...a really bad one. Might not be as bad as I thought.
Winning as O player was causing the final prompt to change to X and was
recording the wrong win. Prompt stays where it should and after extensive
testing it seems the records are correct but just updating on screen
at the wrong time some times. Leaving as is for now.