TIC TAC TOE bugfix5 - finale...
This finally squashes a bug that was bigger than i thought. I had 2
issues...my cat win logic was a little wonky, but just happened to work.
However, i was not resetting the game on sign out which caused
numberOfMoves to stick in memory and come back in a new game.

Now signing out starts a fresh game.

Also, decreases the time of the stats refresh. Feels much more intuitive
and does not apparently break things. testing further but calling this
good for the night.